// product-plan/sections/workspace/types.ts
export type EventTag =
  | 'event'
  | 'raid'
  | 'update'
  | 'community'
  | 'research'
  | 'spotlight'
  | 'misc'